+++ firebase_rules/readme/match.dart
@@ -3,8 +3,8 @@ import 'paths.dart';
 
 @FirebaseRules(service: Service.firestore)
 final firestoreRules = [
-  /// Always start with this match. [FirestorePath] is the root of Firestore.
-  Match<FirestorePath, FirestoreResource>(
+  /// Always start with this match. [FirestoreRoot] is the root of Firestore.
+  Match<FirestoreRoot, FirestoreResource>(
     /// Match statements give access to type-safe contextual information:
     /// - [path] is the path class of this match
     /// - [request] gives access to the [Request] object
@@ -22,8 +22,8 @@ final firestoreRules = [
 
 @FirebaseRules(service: Service.storage)
 final storageRules = [
-  /// Always start with this match. [StoragePath] is the root of Storage.
-  Match<StoragePath, StorageResource>(
+  /// Always start with this match. [StorageRoot] is the root of Storage.
+  Match<StorageRoot, StorageResource>(
     matches: (path, request, resource) => [
       /// All storage matches use [StorageResource] objects
       Match<UsersPath, StorageResource>(),